requirements.txt
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
Flask==1.1.2
Flask-JWT-Extended==3.25.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
requests==2.25.0
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
